+++ packages/graphql/src/schema/questionData.ts
@@ -1,5 +1,5 @@
 import * as DB from '@klicker-uzh/prisma'
-import { BaseQuestionData, DisplayMode } from '@klicker-uzh/util'
+import { type BaseQuestionData, DisplayMode } from '@klicker-uzh/types'
 import builder from '../builder.js'
 
 export const ElementType = builder.enumType('ElementType', {
