+++ packages/types/tsup.config.ts
@@ -1,13 +1,13 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-  entry: ['src/index.ts', 'src/pure.ts', 'src/types.ts'],
+  entry: ['src/index.ts'],
   clean: true,
   format: 'esm',
   target: 'node20',
   dts: true,
   // splitting: false,
-  sourcemap: true,
+  // sourcemap: true,
   // treeshake: true,
   // skipNodeModulesBundle: true,
 })
